Call devfs_init_rulesets() before starting the jails
I'm still in the opinion that an excplicit loading of the devfs rules is
a hack. The loading should be done by another rc script (in fact the
"devfs" script does this, but only if the user enables it in rc.conf) or
jail(8). While a patch for jail(8) was held out in prospect, nothing
happend for about 5 month. Therefor solve this issue by ourself.